requirements.txt
torch==2.3.1
torchvision==0.18.1
efficientnet_pytorch==0.7.1
numpy>=1.23.5

mlflow-skinny>=2.9.2
pandas>=1.5.0

pyspark==3.5.0
dataclasses==0.6

Pillow>=9.0.1
pyarrow
packaging
importlib-metadata<=7.1
hyperopt==0.2.5
petastorm==0.12.0
